requirements.txt


langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langchain-chroma>=0.1.4,<0.2.0
langchain-text-splitters>=0.3.0,<0.4.0


chromadb>=0.5.0,<0.6.0


openai>=1.40.0,<2.0.0


unstructured[pdf]>=0.15.0,<0.16.0

python-dotenv>=1.0.0        
pydantic>=2.0.0              
tiktoken>=0.7.0               


pytest>=8.0.0
pytest-asyncio>=0.23.0
